+++ app/Models/Action.php
@@ -7,7 +7,7 @@ use Illuminate\Database\Eloquent\Model;
 class Action extends Model
 {
     protected $guarded = ['id', 'created_at', 'updated_at'];
-    protected $fillable = ['name', 'brand_id', 'city_id', 'active_from', 'active_to', 'text', 'upload_id', 'status', 'links', 'type'];
+    //protected $fillable = ['name', 'brand_id', 'city_id', 'active_from', 'active_to', 'text', 'upload_id', 'status', 'links', 'type'];
         
     public function categories()
     {
