+++ app/Models/Brand.php
@@ -7,7 +7,7 @@ use Illuminate\Database\Eloquent\Model;
 class Brand extends Model
 {
     protected $guarded = ['id', 'created_at', 'updated_at'];
-    protected $fillable = ['name', 'city_id', 'address', 'phone', 'upload_id', 'links'];
+    //protected $fillable = ['name', 'city_id', 'address', 'phone', 'upload_id', 'links'];
     
     public function uploads()
     {
