+++ app/Models/Upload.php
@@ -7,7 +7,7 @@ use Illuminate\Database\Eloquent\Model;
 class Upload extends Model
 {
     protected $guarded = ['id', 'created_at', 'updated_at'];
-    protected $fillable = ['path', 'ext', 'size', 'old_name'];
+    //protected $fillable = ['path', 'ext', 'size', 'old_name'];
     
     public function actions()
     {
