+++ app/Models/User.php
@@ -9,7 +9,7 @@ use Illuminate\Foundation\Auth\User as Authenticatable;
 class User extends Authenticatable
 {
 	protected $guarded = ['id', 'created_at', 'updated_at'];
-    protected $fillable = ['name', 'email', 'email_verified_at', 'password'];
+    //protected $fillable = ['name', 'email', 'email_verified_at', 'password'];
     protected $hidden = ['password', 'remember_token'];
     
     public function roles()
